Change mixer to float64, add DynamicRangeCompressor

+++ compressor/Compressor.go
@@ -2,5 +2,5 @@ package compressor
 
 //Compressor interface for sound compressor's
 type Compressor interface {
-	Compress(notCompressedSound int) int
+	Compress(notCompressedSound float64) float64
 }
